requirements.txt
alembic==1.4.3
APScheduler==3.6.3
beautifulsoup4==4.9.3
blinker==1.4
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dominate==2.5.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-APScheduler==1.11.0
Flask-Avatars==0.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-OAuthlib==0.9.6
Flask-Share==0.1.1
Flask-SQLAlchemy==2.4.4
flask-whooshee==0.7.0
Flask-WTF==0.14.3
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
oauthlib==2.1.0
Pillow==7.2.0
psutil==5.8.0
PyExecJS==1.5.1
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.1.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.19
tzlocal==2.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.7.4
wordcloud==1.8.0
WTForms==2.3.3
